Fixed MVC Pattern
Fixed MVC pattern.

Add BombsController and CharactersController.

+++ Bomb/Assets/Scripts/Interfaces/IDamage.cs
@@ -2,6 +2,6 @@
 {
     public interface IDamage
     {
-        void TakeDamage(float value);
+        void TakeDamage(int value);
     }
 }
